add "vespa-start-java" helper
